backend/jenkins/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
croniter==6.0.0
Django==5.2.7
djangorestframework==3.16.1
dramatiq==1.18.0
idna==3.10
kombu==5.5.4
packaging==25.0
pika==1.3.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
